# docs/openapi.yaml
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/categorias/{category}": {
            "get": {
                "operationId": "categorias.show",
                "summary": "Display the specified resource",
                "tags": [
                    "Categoria"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "description": "The category ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`CategoriaResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/CategoriaResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "productos"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            },
            "put": {
                "operationId": "categorias.update",
                "summary": "Update the specified resource in storage",
                "tags": [
                    "Categoria"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "description": "The category ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoriaUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`CategoriaResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CategoriaResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "categorias.destroy",
                "summary": "Remove the specified resource from storage",
                "tags": [
                    "Categoria"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "description": "The category ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/comandas/{comanda}": {
            "get": {
                "operationId": "comandas.show",
                "summary": "Display the specified resource",
                "tags": [
                    "Comanda"
                ],
                "parameters": [
                    {
                        "name": "comanda",
                        "in": "path",
                        "required": true,
                        "description": "The comanda ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`ComandaResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/ComandaResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "mesa",
                                                        "mesero",
                                                        "estado",
                                                        "detalles"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            },
            "put": {
                "operationId": "comandas.update",
                "description": "Si el cambio de estado corresponde al estado \"cerrada\", se delega en\nComandaService::cerrar() para aplicar la regla de negocio (no cerrar sin\ndetalle, c\u00e1lculo del descuento y liberaci\u00f3n de la mesa). Cualquier otro\ncambio se aplica directamente.",
                "summary": "Update the specified resource in storage",
                "tags": [
                    "Comanda"
                ],
                "parameters": [
                    {
                        "name": "comanda",
                        "in": "path",
                        "required": true,
                        "description": "The comanda ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ComandaUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ComandaResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/ComandaResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "mesa",
                                                        "mesero",
                                                        "estado",
                                                        "detalles"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "comandas.destroy",
                "summary": "Remove the specified resource from storage",
                "tags": [
                    "Comanda"
                ],
                "parameters": [
                    {
                        "name": "comanda",
                        "in": "path",
                        "required": true,
                        "description": "The comanda ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/detalles/{detalleComanda}": {
            "get": {
                "operationId": "detalles.show",
                "description": "Ruta \"shallow\": /detalles/{detalleComanda}, sin repetir la comanda.",
                "summary": "Display the specified resource",
                "tags": [
                    "DetalleComanda"
                ],
                "parameters": [
                    {
                        "name": "detalleComanda",
                        "in": "path",
                        "required": true,
                        "description": "The detalle comanda ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`DetalleComandaResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/DetalleComandaResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "producto"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            },
            "put": {
                "operationId": "detalles.update",
                "summary": "Update the specified resource in storage",
                "tags": [
                    "DetalleComanda"
                ],
                "parameters": [
                    {
                        "name": "detalleComanda",
                        "in": "path",
                        "required": true,
                        "description": "The detalle comanda ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DetalleComandaUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`DetalleComandaResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/DetalleComandaResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "producto"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "detalles.destroy",
                "summary": "Remove the specified resource from storage",
                "tags": [
                    "DetalleComanda"
                ],
                "parameters": [
                    {
                        "name": "detalleComanda",
                        "in": "path",
                        "required": true,
                        "description": "The detalle comanda ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/estados-comanda/{estadoComanda}": {
            "get": {
                "operationId": "estados-comanda.show",
                "summary": "Display the specified resource",
                "tags": [
                    "EstadoComanda"
                ],
                "parameters": [
                    {
                        "name": "estadoComanda",
                        "in": "path",
                        "required": true,
                        "description": "The estado comanda ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`EstadoComandaResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EstadoComandaResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            },
            "put": {
                "operationId": "estados-comanda.update",
                "summary": "Update the specified resource in storage",
                "tags": [
                    "EstadoComanda"
                ],
                "parameters": [
                    {
                        "name": "estadoComanda",
                        "in": "path",
                        "required": true,
                        "description": "The estado comanda ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstadoComandaUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`EstadoComandaResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EstadoComandaResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "estados-comanda.destroy",
                "summary": "Remove the specified resource from storage",
                "tags": [
                    "EstadoComanda"
                ],
                "parameters": [
                    {
                        "name": "estadoComanda",
                        "in": "path",
                        "required": true,
                        "description": "The estado comanda ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/mesas/{mesa}": {
            "get": {
                "operationId": "mesas.show",
                "summary": "Display the specified resource",
                "tags": [
                    "Mesa"
                ],
                "parameters": [
                    {
                        "name": "mesa",
                        "in": "path",
                        "required": true,
                        "description": "The mesa ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`MesaResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/MesaResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "estado"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            },
            "put": {
                "operationId": "mesas.update",
                "summary": "Update the specified resource in storage",
                "tags": [
                    "Mesa"
                ],
                "parameters": [
                    {
                        "name": "mesa",
                        "in": "path",
                        "required": true,
                        "description": "The mesa ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MesaUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`MesaResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/MesaResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "estado"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "mesas.destroy",
                "summary": "Remove the specified resource from storage",
                "tags": [
                    "Mesa"
                ],
                "parameters": [
                    {
                        "name": "mesa",
                        "in": "path",
                        "required": true,
                        "description": "The mesa ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CategoriaResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "productos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductoResource"
                        }
                    },
                    "creado_en": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "actualizado_en": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "nombre",
                    "descripcion",
                    "creado_en",
                    "actualizado_en"
                ],
                "title": "CategoriaResource"
            },
...
            "CategoriaUpdateRequest": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "descripcion": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "title": "CategoriaUpdateRequest"
            },
...
            "ComandaUpdateRequest": {
                "type": "object",
                "properties": {
                    "mesa_id": {
                        "type": "integer"
                    },
                    "estado_comanda_id": {
                        "type": "integer"
                    },
                    "cerrada_en": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                },
                "title": "ComandaUpdateRequest"
            },
...
            "DetalleComandaUpdateRequest": {
                "type": "object",
                "properties": {
                    "cantidad": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "notas": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    }
                },
                "title": "DetalleComandaUpdateRequest"
            },
...
        },
        "responses": {
...
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}